requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
boto3==1.35.66; python_version >= '3.8'
boto3-stubs==1.35.66; python_version >= '3.8'
botocore==1.35.66; python_version >= '3.8'
botocore-stubs==1.35.66; python_version >= '3.8'
django==4.2.16; python_version >= '3.8'
django-storages==1.14.4; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
pillow==11.0.0; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
s3transfer==0.10.4; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.2; python_version >= '3.8'
types-awscrt==0.23.0; python_version >= '3.8'
types-s3transfer==0.10.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.12'
urllib3==1.26.20; python_version < '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
